Checkpoint. TODO split this into multiple commits
Remove the potentially returning of usize from section_number/section_string
Rather than trust the function to return offset, I can use the outer
offset function.

Convert to a trait for Section because I intended to make a
SectionOfSections, but I'll probably convert to BufferedReader soon
anyway and not need it.

Update validation example to be actually more useful. Currently uses
json crate to validate strings, but I should implement that eventually.

I'd also like to be able to have an option to get a next_token which
also decodes the String rather than finding the edges.

Add check_next method which cleans up the function.

Remove any method which requires backtracking so that I can be ready to
use BufferedReader with sections of discontinuous data.

Make json-extractor and have it respect the number boundary edge case.